backend/requirements.txt
# SmartLib Kiosk - Core System
# Face Recognition + Book Detection + Transaction Management

# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator>=2.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.3
pgvector>=0.2.0

# AI/ML - Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Face Recognition (Deep Neural Networks)
insightface==0.7.3
# onnxruntime-gpu>=1.16.0
onnxruntime>=1.16.0  # Use this for CPU-only

# Object Detection (YOLOv8)
ultralytics>=8.0.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Barcode
pyzbar>=0.1.9

# Image Processing
scikit-image>=0.21.0
albumentations>=1.3.0

# API & Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Logging
loguru>=0.7.0
